refactor common files
