+++ src/test/resources/ieee14-macroconnects/IEEE14.jobs
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <dyn:jobs xmlns:dyn="http://www.rte-france.com/dynawo">
     <dyn:job name="Job">
-        <dyn:solver lib="dynawo_SolverIDA" parFile="ieee14-macroconnects-solvers.par" parId="2"/>
+        <dyn:solver lib="dynawo_SolverIDA" parFile="ieee-solvers.par" parId="2"/>
         <dyn:modeler compileDir="outputs/compilation">
             <dyn:network iidmFile="powsybl_dynawo.xiidm" parFile="ieee14-macroconnects-network.par" parId="8"/>
             <dyn:dynModels dydFile="powsybl_dynawo.dyd"/>
