+++ src/test/resources/ieee14-macroconnects/IEEE14.par
@@ -1,16 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!--
-        Copyright (c) 2015-2019, RTE (http://www.rte-france.com)
-        See AUTHORS.txt
-        All rights reserved.
-        This Source Code Form is subject to the terms of the Mozilla Public
-        License, v. 2.0. If a copy of the MPL was not distributed with this
-        file, you can obtain one at http://mozilla.org/MPL/2.0/.
-        SPDX-License-Identifier: MPL-2.0
-
-        This file is part of Dynawo, an hybrid C++/Modelica open source time domain
-        simulation tool for power systems.
--->
 <parametersSet xmlns="http://www.rte-france.com/dynawo">
     <set id="LAB">
         <par type="DOUBLE" name="load_alpha" value="1.5"/>
